Move parsed types to their own module
This will allow them to be enabled only by a particular feature so that
dependents can choose whether to depend on the `indexmap` crate.
